// webclient/src/gen/message.ts
export enum FileType {
  Dir = 0,
  DirLink = 2,
  DirDrive = 3,
  File = 4,
  FileLink = 5,
}
// ...
export interface FileEntry {
  entry_type: FileType;
  name: string;
  is_hidden: boolean;
  size: bigint;
  modified_time: bigint;
}

export interface FileDirectory {
  id: number;
  path: string;
  entries: FileEntry[];
}

export interface ReadDir {
  path: string;
  include_hidden: boolean;
}

export interface ReadEmptyDirs {
  path: string;
  include_hidden: boolean;
}

export interface ReadEmptyDirsResponse {
  path: string;
  empty_dirs: FileDirectory[];
}
// ...
export interface FileResponse {
  union?:
    | { $case: "dir"; dir: FileDirectory }
    | { $case: "block"; block: FileTransferBlock }
    | { $case: "error"; error: FileTransferError }
    | { $case: "done"; done: FileTransferDone }
    | { $case: "digest"; digest: FileTransferDigest }
    | { $case: "empty_dirs"; empty_dirs: ReadEmptyDirsResponse }
    | undefined;
}

export interface FileTransferDigest {
  id: number;
  file_num: number;
  last_modified: bigint;
  file_size: bigint;
  is_upload: boolean;
  is_identical: boolean;
  /** For resume. Indicates the size of the file already transferred */
  transferred_size: bigint;
  /** For resume. Indicates if the transfer is a resume. */
  is_resume: boolean;
}

export interface FileTransferBlock {
  id: number;
  file_num: number;
  data: Uint8Array;
  compressed: boolean;
  blk_id: number;
}

export interface FileTransferError {
  id: number;
  error: string;
  file_num: number;
}
// ...
export interface FileTransferDone {
  id: number;
  file_num: number;
}
// ...
export const FileEntry: MessageFns<FileEntry> = {
  encode(message: FileEntry, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.entry_type !== 0) {
      writer.uint32(8).int32(message.entry_type);
    }
    if (message.name !== "") {
      writer.uint32(18).string(message.name);
    }
    if (message.is_hidden !== false) {
      writer.uint32(24).bool(message.is_hidden);
    }
    if (message.size !== 0n) {
      if (BigInt.asUintN(64, message.size) !== message.size) {
        throw new globalThis.Error("value provided for field message.size of type uint64 too large");
      }
      writer.uint32(32).uint64(message.size);
    }
    if (message.modified_time !== 0n) {
      if (BigInt.asUintN(64, message.modified_time) !== message.modified_time) {
        throw new globalThis.Error("value provided for field message.modified_time of type uint64 too large");
      }
      writer.uint32(40).uint64(message.modified_time);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FileEntry {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFileEntry();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.entry_type = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.is_hidden = reader.bool();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.size = reader.uint64() as bigint;
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.modified_time = reader.uint64() as bigint;
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<FileEntry>, I>>(base?: I): FileEntry {
    return FileEntry.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<FileEntry>, I>>(object: I): FileEntry {
    const message = createBaseFileEntry();
    message.entry_type = object.entry_type ?? 0;
    message.name = object.name ?? "";
    message.is_hidden = object.is_hidden ?? false;
    message.size = (object.size !== undefined && object.size !== null) ? BigInt(object.size) : 0n;
    message.modified_time = (object.modified_time !== undefined && object.modified_time !== null)
      ? BigInt(object.modified_time)
      : 0n;
    return message;
  },
};

function createBaseFileDirectory(): FileDirectory {
  return { id: 0, path: "", entries: [] };
}

export const FileDirectory: MessageFns<FileDirectory> = {
  encode(message: FileDirectory, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== 0) {
      writer.uint32(8).int32(message.id);
    }
    if (message.path !== "") {
      writer.uint32(18).string(message.path);
    }
    for (const v of message.entries) {
      FileEntry.encode(v!, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FileDirectory {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFileDirectory();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.id = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.path = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.entries.push(FileEntry.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<FileDirectory>, I>>(base?: I): FileDirectory {
    return FileDirectory.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<FileDirectory>, I>>(object: I): FileDirectory {
    const message = createBaseFileDirectory();
    message.id = object.id ?? 0;
    message.path = object.path ?? "";
    message.entries = object.entries?.map((e) => FileEntry.fromPartial(e)) || [];
    return message;
  },
};
// ...
export const FileResponse: MessageFns<FileResponse> = {
  encode(message: FileResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    switch (message.union?.$case) {
      case "dir":
        FileDirectory.encode(message.union.dir, writer.uint32(10).fork()).join();
        break;
      case "block":
        FileTransferBlock.encode(message.union.block, writer.uint32(18).fork()).join();
        break;
      case "error":
        FileTransferError.encode(message.union.error, writer.uint32(26).fork()).join();
        break;
      case "done":
        FileTransferDone.encode(message.union.done, writer.uint32(34).fork()).join();
        break;
      case "digest":
        FileTransferDigest.encode(message.union.digest, writer.uint32(42).fork()).join();
        break;
      case "empty_dirs":
        ReadEmptyDirsResponse.encode(message.union.empty_dirs, writer.uint32(50).fork()).join();
        break;
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FileResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFileResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.union = { $case: "dir", dir: FileDirectory.decode(reader, reader.uint32()) };
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.union = { $case: "block", block: FileTransferBlock.decode(reader, reader.uint32()) };
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.union = { $case: "error", error: FileTransferError.decode(reader, reader.uint32()) };
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.union = { $case: "done", done: FileTransferDone.decode(reader, reader.uint32()) };
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.union = { $case: "digest", digest: FileTransferDigest.decode(reader, reader.uint32()) };
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.union = { $case: "empty_dirs", empty_dirs: ReadEmptyDirsResponse.decode(reader, reader.uint32()) };
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<FileResponse>, I>>(base?: I): FileResponse {
    return FileResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<FileResponse>, I>>(object: I): FileResponse {
    const message = createBaseFileResponse();
    switch (object.union?.$case) {
      case "dir": {
        if (object.union?.dir !== undefined && object.union?.dir !== null) {
          message.union = { $case: "dir", dir: FileDirectory.fromPartial(object.union.dir) };
        }
        break;
      }
      case "block": {
        if (object.union?.block !== undefined && object.union?.block !== null) {
          message.union = { $case: "block", block: FileTransferBlock.fromPartial(object.union.block) };
        }
        break;
      }
      case "error": {
        if (object.union?.error !== undefined && object.union?.error !== null) {
          message.union = { $case: "error", error: FileTransferError.fromPartial(object.union.error) };
        }
        break;
      }
      case "done": {
        if (object.union?.done !== undefined && object.union?.done !== null) {
          message.union = { $case: "done", done: FileTransferDone.fromPartial(object.union.done) };
        }
        break;
      }
      case "digest": {
        if (object.union?.digest !== undefined && object.union?.digest !== null) {
          message.union = { $case: "digest", digest: FileTransferDigest.fromPartial(object.union.digest) };
        }
        break;
      }
      case "empty_dirs": {
        if (object.union?.empty_dirs !== undefined && object.union?.empty_dirs !== null) {
          message.union = {
            $case: "empty_dirs",
            empty_dirs: ReadEmptyDirsResponse.fromPartial(object.union.empty_dirs),
          };
        }
        break;
      }
    }
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | bigint | undefined;

export type DeepPartial<T> = T extends bigint ? string | number | bigint
  : T extends Builtin ? T : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}